Vectors comparisons

Identify similarities and differences between two R vectors.

vectors_comparisons(first_vector, second_vector, comparison_type, output)

Arguments

first_vector
vector expected.

second_vector
vector expected.

comparison_type
character expected. Type of comparison expected, you can choose between "difference" or "equality". "Difference" highlight element(s) of the first vector not present in the second, and "equality" check if two vectors are exactly the same.

output
character expected. Kind of expected output. You can choose between "message", "report" or "logical".

Examples

vectors_comparisons(first_vector = c(1, 2, 3, 5),
                    second_vector = c(1, 2, 4, 3),
                    comparison_type = "difference",
                    output = "report")
#> # A tibble: 4 × 2
#>   first_vector vectors_comparisons_output           
#>          <dbl> <chr>                                
#> 1            1 no difference                        
#> 2            2 no difference                        
#> 3            3 no difference                        
#> 4            5 data not present in the second sector